+++ src/dispatching/dialogue/storage/mod.rs
@@ -73,3 +73,85 @@ pub trait Storage<D> {
         chat_id: i64,
     ) -> BoxFuture<'static, Result<Option<D>, Self::Error>>;
 }
+
+/// A storage with an erased error type.
+pub type ErasedStorage<D> = Arc<dyn Storage<D, Error = Box<dyn std::error::Error>>>;
+
+/// Extension methods for working with [`Storage`].
+pub trait StorageExt<D> {
+    /// Returns [`ErasedStorage`].
+    fn erase(self: Arc<Self>) -> ErasedStorage<D>;
+}
+
+impl<S, D> StorageExt<D> for S
+where
+    S: Storage<D> + Send + Sync + 'static,
+    S::Error: std::error::Error + 'static,
+{
+    fn erase(self: Arc<Self>) -> ErasedStorage<D> {
+        struct Eraser<S>(Arc<S>);
+
+        impl<D, S> Storage<D> for Eraser<S>
+        where
+            S: Storage<D> + Send + Sync + 'static,
+            S::Error: std::error::Error + 'static,
+        {
+            type Error = Box<dyn std::error::Error>;
+
+            fn remove_dialogue(
+                self: Arc<Self>,
+                chat_id: i64,
+            ) -> BoxFuture<'static, Result<(), Self::Error>>
+            where
+                D: Send + 'static,
+            {
+                Box::pin(async move {
+                    Arc::clone(&self.0).remove_dialogue(chat_id).await.map_err(|e| e.into())
+                })
+            }
+
+            fn update_dialogue(
+                self: Arc<Self>,
+                chat_id: i64,
+                dialogue: D,
+            ) -> BoxFuture<'static, Result<(), Self::Error>>
+            where
+                D: Send + 'static,
+            {
+                Box::pin(async move {
+                    Arc::clone(&self.0)
+                        .update_dialogue(chat_id, dialogue)
+                        .await
+                        .map_err(|e| e.into())
+                })
+            }
+
+            fn get_dialogue(
+                self: Arc<Self>,
+                chat_id: i64,
+            ) -> BoxFuture<'static, Result<Option<D>, Self::Error>> {
+                Box::pin(async move {
+                    Arc::clone(&self.0).get_dialogue(chat_id).await.map_err(|e| e.into())
+                })
+            }
+        }
+
+        Arc::new(Eraser(self))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    async fn test_erased() {
+        let chat_id = 123;
+
+        let erased = InMemStorage::new().erase();
+        Arc::clone(&erased).update_dialogue(chat_id, 1).await.unwrap();
+        assert_eq!(Arc::clone(&erased).get_dialogue(chat_id).await.unwrap().unwrap(), 1);
+        Arc::clone(&erased).remove_dialogue(chat_id).await.unwrap();
+        assert_eq!(Arc::clone(&erased).get_dialogue(chat_id).await.unwrap(), None);
+    }
+}